Extract cloneable repository list out of clone component for reuse
